rename: 공고등록 문의 및 토마토들 소개 컴포넌트 containers로 이전
Related to: #32
