+++ src/containers/cs/Terms.tsx
